requirements.txt
requests
qdrant-client
python-dotenv
pdfplumber
beautifulsoup4
tqdm